requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
python-dotenv
email-validator
passlib==1.7.4
bcrypt==3.2.0
